Fix supported CPU bitmap handling
The previous approach used shift operations that could
exceed the bit-width of the shifted type, leading to
undefined results. Fix this, and also align the cpu selection
mechanism with the standard kernel syntax for setting isolcpus
to unify kernel command line options.

Also use the opportunity to bail out if the user selects a bogus
subset that leaves no CPUs for Xenomai to run on, and make
sure that only CPUs that are present can actually be selected
(otherwise, we run into problems with the new ipipe timer selection
code that will try to find timers for non-existent CPUs)

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>